requirements.txt
python-gitlab=4.4.0
python-dotenv=1.0.0